update dir and name of ser_vi_layoutxlm model
