+++ fastdeploy/vision/ocr/ppocr/structurev2_ser_vi_layoutxlm.h
@@ -28,9 +28,9 @@ class FASTDEPLOY_DECL SERViLayoutXLMModel : public FastDeployModel {
  public:
   /** \brief Set path of model file and configuration file, and the configuration of runtime
    *
-   * \param[in] model_file Path of model file, e.g resnet/model.pdmodel
-   * \param[in] params_file Path of parameter file, e.g resnet/model.pdiparams, if the model format is ONNX, this parameter will be ignored
-   * \param[in] config_file Path of configuration file for deployment, e.g resnet/infer_cfg.yml
+   * \param[in] model_file Path of model file, e.g ser_vi_layoutxlm/model.pdmodel
+   * \param[in] params_file Path of parameter file, e.g ser_vi_layoutxlm/model.pdiparams, if the model format is ONNX, this parameter will be ignored
+   * \param[in] config_file Path of configuration file for deployment, e.g ser_vi_layoutxlm/infer_cfg.yml
    * \param[in] custom_option RuntimeOption for inference, the default will use cpu, and choose the backend defined in `valid_cpu_backends`
    * \param[in] model_format Model format of the loaded model, default is Paddle format
    */
@@ -47,7 +47,7 @@ class FASTDEPLOY_DECL SERViLayoutXLMModel : public FastDeployModel {
   virtual std::unique_ptr<SERViLayoutXLMModel> Clone() const;
 
   /// Get model's name
-  virtual std::string ModelName() const { return "SERViLayoutxlm"; }
+  virtual std::string ModelName() const { return "SERViLayoutXLMModel"; }
 
  protected:
   bool Initialize();
