+++ python/fastdeploy/vision/ocr/__init__.py
@@ -14,4 +14,3 @@
 from __future__ import absolute_import
 
 from .ppocr import *
-from .ser_vi_layoutxlm import *
